Align CLI auth with homepage sign-in
