+++ apps/web/components/auth/auth-guard.tsx
@@ -6,9 +6,11 @@ import { SignInButton } from "./sign-in-button";
 export function AuthGuard({
   children,
   loadingFallback,
+  unauthenticatedFallback,
 }: {
   children: React.ReactNode;
   loadingFallback?: React.ReactNode;
+  unauthenticatedFallback?: React.ReactNode;
 }) {
   const { loading, isAuthenticated } = useSession();
 
@@ -18,10 +20,14 @@ export function AuthGuard({
 
   if (!isAuthenticated) {
     return (
-      <div className="flex flex-col items-center gap-4 p-8">
-        <p>Please sign in to continue</p>
-        <SignInButton />
-      </div>
+      <>
+        {unauthenticatedFallback ?? (
+          <div className="flex flex-col items-center gap-4 p-8">
+            <p>Please sign in to continue</p>
+            <SignInButton />
+          </div>
+        )}
+      </>
     );
   }
